requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aioprocessing==2.0.1
aioresponses==0.7.6
aiosignal==1.3.1
altgraph==0.17.4
analytics-python==1.2.9
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
argcomplete==3.2.1
attrs==23.2.0
backoff==2.2.1
backstage==0.0.22
beautifulsoup4==4.9.3
blinker==1.8.1
boto3==1.34.98
botocore==1.34.98
braq==0.0.6
Brotli==1.1.0
cement==2.6.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
distro==1.9.0
dnspython==2.6.1
docker-pycreds==0.4.0
-e git+https://github.com/droope/droopescan.git@7d61a537e2c2f9b794c0ce7f94f6a5c4f059bbd4#egg=droopescan
EditorConfig==0.12.4
email_validator==2.1.1
et-xmlfile==1.1.0
fastapi==0.111.0
fastapi-cli==0.0.2
fastapi-sso==0.14.2
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
gql==3.5.0
graphql-core==3.2.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpretty==1.1.4
httptools==0.6.1
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
itsdangerous==2.2.0
janus==1.0.0
jesth==0.0.11
Jinja2==3.1.3
jmespath==1.0.1
jsbeautifier==1.15.1
kaitaistruct==0.10
LinkFinder==1.0
lxml==5.2.0
macholib==1.16.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
menubar==0.0.1
modulegraph==0.19.6
multidict==6.0.5
ninja==1.11.1.1
Nuitka==2.2.3
numpy @ file:///private/tmp/numpy-20240206-5894-1wazl8/numpy-1.26.4
oauthlib==3.2.2
ollama==0.1.9
openai==1.25.1
openpyxl==3.1.2
openui @ file:///Users/bnktop/Desktop/Projects/h/openui/backend
ordered-set==4.1.0
orjson==3.10.3
oscan==0.0.6
outcome==1.3.0.post0
packaging==23.2
paradict==0.0.7
pdfkit==1.0.0
peewee==3.17.3
phonenumbers==8.12.25
pillow @ file:///private/tmp/pillow-20240123-12080-wk3cb3/pillow-10.2.0
platformdirs==4.1.0
protobuf==4.25.3
psutil==5.9.8
py2app==0.28.7
pyarrow==16.0.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
pyobjc-core==10.2
pyobjc-framework-Cocoa==10.2
pyobjc-framework-Quartz==10.2
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyQRCode==1.2.1
PySocks==1.7.1
pystache==0.6.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-Wappalyzer==0.3.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rumps==0.4.0
s3transfer==0.10.1
selenium==4.20.0
selenium-wire==5.1.0
sentry-sdk==2.0.1
setproctitle==1.3.3
shared==0.0.30
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.37.2
stem==1.8.2
subrun==0.0.8
tiktoken==0.6.0
tld==0.13
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.11.0
ujson==5.9.0
urllib3==2.2.1
userpath==1.9.1
ustrid==0.0.3
uvicorn==0.29.0
uvloop==0.19.0
wandb==0.16.6
watchfiles==0.21.0
weave @ git+https://github.com/wandb/weave.git@bf3ecff27b4809ba6d189ab729b04fad98eee973
websockets==12.0
Werkzeug==3.0.2
written==0.0.4
wsproto==1.2.0
yarl==1.9.4
zstandard==0.22.0
